Restore upstream scoring heuristics on GPU brute force; auto CPU-lite fallback
Scoring (ZScore, GPU + CPU identical formula):
- Restore order boost (non-linear 1/(pos+1)^strength), legendary stat
  multiplier (default 2x), and overcap leftover points dropped by the
  v0.1.7.3 rebase simplification
- Weights are pre-combined on CPU (BuildStatConfig) and passed to the
  HLSL kernel as float StatMul; ZScore branch now snaps to breakpoint
  levels and adds unweighted overcap points
- New settings sliders: Priority Order Boost / Legendary Stat Multiplier
  (ZScore only, disabled in Combat Power mode)

Compute backend:
- Remove the GPU/CPU toggle (SolverConfig.UseGpu); GPU is always
  preferred and failures fall back automatically (resets on restart)
- CPU fallback now uses the lightweight beam search instead of the
  branch-and-bound path: bounded runtime independent of combo count,
  same score formula as the GPU (ZScore and Combat Power both)
- Results panel shows an "approximate (CPU-lite)" indicator via
  SolverResult.UsedCpuFallback

Beam search fixes:
- ModuleSetIndices fixed buffer 5 -> MaxModules(10): prevents unsafe
  out-of-bounds writes when solving >5 module sets on the fallback path
- StatDifference: Math.Abs(statA = statB) -> statA - statB (upstream
  typo that disabled result dedup)
- CombatPower support: per-stat FightValue lookup + global
  LinkTotalFight term, mirroring CalcCombosCombatScore
- BuildStatScoreLookup resolves original stat ids via PossibleStats
  (fixes wrong-stat lookup when priorities are missing from inventory)

Localization: replace Module_Settings_UseGpu with GpuPending, add
OrderBoostStrength/LegendaryMul labels+tooltips and Module_Result_CpuFallback
(ext.en/ext.ja). Update README-JP.md and FORK_NOTES.md accordingly.

+++ BPSR-ZDPS/Managers/Gpu/GpuComputeContext.cs
@@ -32,7 +32,7 @@ namespace BPSR_ZDPS.Managers.Gpu
         public int MaxTotal;            // K * 20
 
         public uint[] ModuleStats = [];   // [N*S]
-        public int[] StatMul = [];        // [S]
+        public float[] StatMul = [];      // [S] combined weight: legendaryMul x orderBoost (0 = ignore)
         public int[] StatReq = [];        // [S]
         public int[] StatMin = [];        // [S]
         public int[] StatExact = [];      // [S]
